Lattice class
Finalized how indexing works. Lattice_vecs is a 3d array, where each 2d matrix slice represents the locations of all the sites for a particular particle in the unit cell. Similarly, spin_matrix is a 3d array where each row in each 2d slice represents the xyz spin of a site of one particular particle in the unit cell.